chore: finishing touches for demo
